+++ site/src/tutorial.rs
@@ -373,8 +373,8 @@ fn TutorialArrays() -> impl IntoView {
         <p>"By using "<Prim prim=Constant/>", we can turn any value into a function that pushes that value onto the stack. We can then put these functions into an array like any other."</p>
         <Editor example="[□1 □2 □[7 8 9]]"/>
         <p>"The "<code>"⟦⟧"</code>"s indicate that a list is wrapped in a constant function."</p>
-        <p>"To get the values back on the stack, we can use "<Prim prim=Reduce/><Prim prim=Call/>"."</p>
-        <Editor example="/![□1 □2 □[7 8 9]]"/>
+        <p>"To get the values back on the stack, we can use "<Prim prim=Reduce/><Prim prim=Invert/><Prim prim=Constant/>"."</p>
+        <Editor example="/⍘□[□1 □2 □[7 8 9]]"/>
         <p>"Having to write "<Prim prim=Constant glyph_only=true/>" everywhere is annoying, and so..."</p>
 
         <h2 id="nested-arrays">"Nested Arrays"</h2>